+++ ZoeSleep/ZoeSleep/Managers/AuthenticationManager.swift
@@ -270,7 +270,7 @@ class AuthenticationManager: ObservableObject {
         let inputData = Data(input.utf8)
         let hashedData = SHA256.hash(data: inputData)
         let hashString = hashedData.compactMap {
-            String(format: "%02, x", $0)
+            String(format: "%02x", $0)
         }.joined()
         return hashString
     }